Remove cli/lib.rs, deno is a "bin" crate only
Because the tests sometimes reached into the crate, some code
duplication was necessary.
