+++ cli/Cargo.toml
@@ -8,15 +8,6 @@ authors = ["the Deno authors"]
 edition = "2018"
 description = "Provides the deno executable"
 repository = "https://github.com/denoland/deno"
-default-run = "deno"
-
-[lib]
-name = "deno"
-path = "lib.rs"
-
-[[bin]]
-name = "deno"
-path = "main.rs"
 
 [build-dependencies]
 deno_core = { path = "../core", version = "0.43.0" }
